// app/xmlns/www.musicxml.org.d.ts
interface BaseType {
	_exists: boolean;
	_namespace: string;
}
// ...
export interface attributes extends _attributes { constructor: { new(): attributes }; }
// ...
export interface backup extends _backup { constructor: { new(): backup }; }
// ...
export interface barline extends _barline { constructor: { new(): barline }; }
// ...
export interface bookmark extends _bookmark { constructor: { new(): bookmark }; }
// ...
export interface direction extends _direction { constructor: { new(): direction }; }
// ...
export interface figuredBass extends _figuredBass { constructor: { new(): figuredBass }; }
// ...
export interface forward extends _forward { constructor: { new(): forward }; }
// ...
export interface grouping extends _grouping { constructor: { new(): grouping }; }
// ...
export interface harmony extends _harmony { constructor: { new(): harmony }; }
// ...
export interface link extends _link { constructor: { new(): link }; }
// ...
export interface note extends _note { constructor: { new(): note }; }
// ...
export interface print extends _print { constructor: { new(): print }; }
// ...
interface _ScorePartwiseType extends BaseType {
	version: string;
	credit?: credit[];
	defaults?: defaults;
	identification?: identification;
	/** The movement-number element specifies the number of a movement. */
	movementNumber?: string;
	/** The movement-title element specifies the title of a movement, not including its number. */
	movementTitle?: string;
	part: ScorePartwiseTypePartType[];
	partList: partList;
	work?: work;
}
export interface ScorePartwiseType extends _ScorePartwiseType { constructor: { new(): ScorePartwiseType }; }
export var ScorePartwiseType: { new(): ScorePartwiseType};

interface _ScorePartwiseTypePartType extends BaseType {
	id: string;
	measure: ScorePartwiseTypePartTypeMeasureType[];
}
export interface ScorePartwiseTypePartType extends _ScorePartwiseTypePartType { constructor: { new(): ScorePartwiseTypePartType }; }
export var ScorePartwiseTypePartType: {new ():ScorePartwiseTypePartType};

interface _ScorePartwiseTypePartTypeMeasureType extends BaseType {
	implicit: yesNo;
	nonControlling: yesNo;
	number: string;
	width: number;
	attributes?: attributes[];
	backup?: backup[];
	barline?: barline[];
	bookmark?: bookmark[];
	direction?: direction[];
	figuredBass?: figuredBass[];
	forward?: forward[];
	grouping?: grouping[];
	harmony?: harmony[];
	link?: link[];
	note?: note[];
	print?: print[];
	sound?: sound[];
}
export interface ScorePartwiseTypePartTypeMeasureType extends _ScorePartwiseTypePartTypeMeasureType { constructor: { new(): ScorePartwiseTypePartTypeMeasureType }; }
// OWENW: Manually exported this for convenience
export var measure: { new(): ScorePartwiseTypePartTypeMeasureType };


interface _ScoreTimewiseType extends BaseType {
	version: string;
	credit?: credit[];
	defaults?: defaults;
	identification?: identification;
	measure: ScoreTimewiseTypeMeasureType[];
	/** The movement-number element specifies the number of a movement. */
	movementNumber?: string;
	/** The movement-title element specifies the title of a movement, not including its number. */
	movementTitle?: string;
	partList: partList;
	work?: work;
}
interface ScoreTimewiseType extends _ScoreTimewiseType { constructor: { new(): ScoreTimewiseType }; }
// ...
export interface sound extends _sound { constructor: { new(): sound }; }
// ...
export type yesNo = ("yes" | "no");
// ...
export interface document extends BaseType {
	/** The score-partwise element is the root element for a partwise MusicXML score. It includes a score-header group followed by a series of parts with measures inside. The document-attributes attribute group includes the version attribute. */
	scorePartwise: ScorePartwiseType;
	/** The score-timewise element is the root element for a timewise MusicXML score. It includes a score-header group followed by a series of measures with parts inside. The document-attributes attribute group includes the version attribute. */
	scoreTimewise: ScoreTimewiseType;
}